feat: implement command palette and refactor search command menu
- Added CommandPaletteHost component to manage global command palette functionality.
- Refactored SearchCommandMenu to utilize the new command palette, removing redundant logic.
- Updated PropertiesPanel to streamline tab content styling.
- Enhanced CSS with Framer-inspired token system for improved design consistency.
- Modified greeting logic in the main page for cleaner code.
- Updated screenshot for the homescreen preview.
